second commit

+++ level20/homework/homework5.py
@@ -1,6 +1,6 @@
 # 5) Write a program that takes a number as input and prints whether it is positive, negative, or zero.
 
-num=int(input("Enter number:"))
+num=float(input("Enter number:"))
 
 if num > 0:
     print("positive")
